+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/bean/entity/HibernateRole.java
@@ -21,10 +21,6 @@ public class HibernateRole implements Bean {
     @Column(name = "id", length = Constraints.LENGTH_ID, nullable = false, unique = true)
     private String stringId;
 
-    // -----------------------------------------------------------外键-----------------------------------------------------------
-    @Column(name = "group_id", length = Constraints.LENGTH_ID)
-    private String groupStringId;
-
     // -----------------------------------------------------------主属性字段-----------------------------------------------------------
     @Column(name = "name", length = Constraints.LENGTH_NAME)
     private String name;
@@ -34,13 +30,6 @@ public class HibernateRole implements Bean {
 
     @Column(name = "remark", length = Constraints.LENGTH_REMARK)
     private String remark;
-
-    // -----------------------------------------------------------多对一-----------------------------------------------------------
-    @ManyToOne(targetEntity = HibernateRoleGroup.class)
-    @JoinColumns({ //
-            @JoinColumn(name = "group_id", referencedColumnName = "id", insertable = false, updatable = false), //
-    })
-    private HibernateRoleGroup group;
 
     // -----------------------------------------------------------一对多-----------------------------------------------------------
     @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernatePexp.class, mappedBy = "role")
@@ -58,6 +47,7 @@ public class HibernateRole implements Bean {
     public HibernateRole() {
     }
 
+    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
     public HibernateStringIdKey getKey() {
         return Optional.ofNullable(stringId).map(HibernateStringIdKey::new).orElse(null);
     }
@@ -66,28 +56,13 @@ public class HibernateRole implements Bean {
         this.stringId = Optional.ofNullable(uuidKey).map(HibernateStringIdKey::getStringId).orElse(null);
     }
 
+    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
     public String getStringId() {
         return stringId;
     }
 
     public void setStringId(String stringId) {
         this.stringId = stringId;
-    }
-
-    public HibernateStringIdKey getGroupKey() {
-        return Optional.ofNullable(groupStringId).map(HibernateStringIdKey::new).orElse(null);
-    }
-
-    public void setGroupKey(HibernateStringIdKey stringIdKey) {
-        this.groupStringId = Optional.ofNullable(stringIdKey).map(HibernateStringIdKey::getStringId).orElse(null);
-    }
-
-    public String getGroupStringId() {
-        return groupStringId;
-    }
-
-    public void setGroupStringId(String groupStringId) {
-        this.groupStringId = groupStringId;
     }
 
     public String getName() {
@@ -114,14 +89,6 @@ public class HibernateRole implements Bean {
         this.remark = remark;
     }
 
-    public HibernateRoleGroup getGroup() {
-        return group;
-    }
-
-    public void setGroup(HibernateRoleGroup group) {
-        this.group = group;
-    }
-
     public Set<HibernatePexp> getPexps() {
         return pexps;
     }
@@ -140,12 +107,10 @@ public class HibernateRole implements Bean {
 
     @Override
     public String toString() {
-        return "HibernateRole{" +
-                "stringId='" + stringId + '\'' +
-                ", groupStringId='" + groupStringId + '\'' +
-                ", name='" + name + '\'' +
-                ", enabled=" + enabled +
-                ", remark='" + remark + '\'' +
-                '}';
+        return getClass().getSimpleName() + "(" +
+                "stringId = " + stringId + ", " +
+                "name = " + name + ", " +
+                "enabled = " + enabled + ", " +
+                "remark = " + remark + ")";
     }
 }