+++ rbac-distributed-service-stack/src/main/java/com/dwarfeng/rbacds/stack/bean/entity/Role.java
@@ -11,10 +11,9 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  */
 public class Role implements Entity<StringIdKey> {
 
-    private static final long serialVersionUID = -1496451428523868925L;
+    private static final long serialVersionUID = 3954346769079299113L;
 
     private StringIdKey key;
-    private StringIdKey groupKey;
     private String name;
     private boolean enabled;
     private String remark;
@@ -22,9 +21,8 @@ public class Role implements Entity<StringIdKey> {
     public Role() {
     }
 
-    public Role(StringIdKey key, StringIdKey groupKey, String name, boolean enabled, String remark) {
+    public Role(StringIdKey key, String name, boolean enabled, String remark) {
         this.key = key;
-        this.groupKey = groupKey;
         this.name = name;
         this.enabled = enabled;
         this.remark = remark;
@@ -38,14 +36,6 @@ public class Role implements Entity<StringIdKey> {
     @Override
     public void setKey(StringIdKey key) {
         this.key = key;
-    }
-
-    public StringIdKey getGroupKey() {
-        return groupKey;
-    }
-
-    public void setGroupKey(StringIdKey groupKey) {
-        this.groupKey = groupKey;
     }
 
     public String getName() {
@@ -76,7 +66,6 @@ public class Role implements Entity<StringIdKey> {
     public String toString() {
         return "Role{" +
                 "key=" + key +
-                ", groupKey=" + groupKey +
                 ", name='" + name + '\'' +
                 ", enabled=" + enabled +
                 ", remark='" + remark + '\'' +