+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 #### 功能移除
 
-- (无)
+- 删除实体。
+  - com.dwarfeng.rbacds.stack.bean.entity.RoleGroup。
 
 ---